fix for path ending with \
if the caller ends the path with \ it will be interpreted as an escape for the following ", which will mess up the args[] array
instead the caller can end the path with \. so the array isn't scambled. Then directoryPath gets trimmed of . and \ at the end.

This is mostly useful for Total Commander which passes backslashes at the end of paths when using %P as argument to commands